+++ core/src/configs/mod.rs
@@ -1,4 +1,5 @@
 pub mod auth_config;
+pub mod azure_config;
 pub mod github_config;
 pub mod job_config;
 pub mod llm_config;
@@ -11,6 +12,4 @@ pub mod stripe_config;
 pub mod throttle_config;
 pub mod user_config;
 pub mod worker_config;
-
-pub mod azure_config;
 pub mod email_config;